Implement q_new
The q_new() dynamically allocating memory using the malloc function for
a block of memory equal to the size of struct list_head.
If the memory allocation is successful, the structure is initialized by
calling the INIT_LIST_HEAD macro, which sets the pointers to the
previous and next nodes to point to itself, indicating an empty list.
Finally, if the memory allocation fails, it returns NULL;
otherwise, it returns a pointer to the newly created queue.